requirements.txt
telethon
google-cloud-bigquery
python-dotenv
loguru 
arabic-reshaper
python-bidi